ad.c, adht.c, app_util.c, cipher.c, pcache.c, sockets.c, util.c, keyd.c, sniffer.c, xcommon.c: removed compiler warnings under xcode compiler and removed dependencies on pipemsg.  In ad.c also do not automatically set sock_v4 to sock_v6; in app_util.c, do not try to call astart_main under iOS, and extended to 20s the time to connect; in sniffer.c, now supports management subtypes

+++ src/mgmt/keyd.c
@@ -97,8 +97,8 @@ static void handle_packet (char * message, int msize)
   if (((msize - (kp - message)) != 513) ||
       (*kp != KEY_RSA4096_E65537)) {
     snprintf (alog->b, alog->s,
-              "msize %d - (%p - %p = %zd) =? 513, *kp %d\n",
-              msize, kp, message, kp - message, *kp);
+              "msize %d - (%p - %p = %d) =? 513, *kp %d\n",
+              msize, kp, message, (int)(kp - message), *kp);
     log_print (alog);
     kp = NULL;
     ksize = 0;
